requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.5.0
requests==2.31.0
cryptography==41.0.7